cmd: Move some license functions from helpers.go to licenses.go
